# docs/swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: FootballCheck
  version: '1.0'
  description: The API for the FootballCheck project
paths:
...
  "/api/v1/coaches":
    get:
      tags:
      - Coach
      parameters:
      - name: Surname
        in: query
        schema:
          type: string
      - name: Country
        in: query
        schema:
          type: string
      - name: MinWorkExperience
        in: query
        schema:
          type: integer
          format: int32
      - name: MaxWorkExperience
        in: query
        schema:
          type: integer
          format: int32
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/CoachSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CoachDto"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CoachDto"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CoachDto"
    post:
      security:
      - Bearer: []
      tags:
      - Coach
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CoachBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/CoachBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CoachBaseDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/coaches/{id}":
    put:
      security:
      - Bearer: []
      tags:
      - Coach
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CoachBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/CoachBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CoachBaseDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
      - Bearer: []
      tags:
      - Coach
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Coach
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
        '404':
          description: Not Found
  "/api/v1/players":
    get:
      tags:
      - Player
      parameters:
      - name: ClubName
        in: query
        schema:
          type: string
      - name: Surname
        in: query
        schema:
          type: string
      - name: Country
        in: query
        schema:
          type: string
      - name: MinPrice
        in: query
        schema:
          type: integer
          format: int32
      - name: MaxPrice
        in: query
        schema:
          type: integer
          format: int32
      - name: MinRating
        in: query
        schema:
          type: integer
          format: int32
      - name: MaxRating
        in: query
        schema:
          type: integer
          format: int32
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/PlayerSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/PlayerDto"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/PlayerDto"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/PlayerDto"
    post:
      security:
      - Bearer: []
      tags:
      - Player
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlayerBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/PlayerBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/PlayerBaseDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/players/{id}":
    put:
      security:
      - Bearer: []
      tags:
      - Player
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlayerBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/PlayerBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/PlayerBaseDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
      - Bearer: []
      tags:
      - Player
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Player
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/PlayerDto"
        '404':
          description: Not Found
...
  "/api/v1/squads/{squadId}/players/{playerId}":
    delete:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/squads/{squadId}/coach/{coachId}":
    delete:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: coachId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
        '409':
          description: Conflict
...
components:
  schemas:
...
    CoachBaseDto:
      type: object
      properties:
        surname:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        workExperience:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    CoachDto:
      type: object
      properties:
        surname:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        workExperience:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
...
    CoachSortState:
      enum:
      - IdAsc
      - IdDesc
      - SurnameAsc
      - SurnameDesc
      - CountryAsc
      - CountryDesc
      - WorkExperienceAsc
      - WorkExperienceDesc
      type: string
...
    PlayerBaseDto:
      type: object
      properties:
        clubId:
          type: integer
          format: int32
          nullable: true
        surname:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
          nullable: true
        country:
          type: string
          nullable: true
        price:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    PlayerDto:
      type: object
      properties:
        clubId:
          type: integer
          format: int32
          nullable: true
        surname:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
          nullable: true
        country:
          type: string
          nullable: true
        price:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
...
    PlayerSortState:
      enum:
      - IdAsc
      - IdDesc
      - SurnameAsc
      - SurnameDesc
      - RatingAsc
      - RatingDesc
      - CountryAsc
      - CountryDesc
      - ClubNameAsc
      - ClubNameDesc
      - PriceAsc
      - PriceDesc
      type: string
...
    SquadDto:
      type: object
      properties:
        coachId:
          type: integer
          format: int32
          nullable: true
        name:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: Please insert JWT with Bearer into field
      name: Authorization
      in: header